Backend/requirements.txt
Flask==2.0.2
werkzeug==2.0.2
sqlalchemy==1.4.*
python-dotenv
mesa==0.9.0
matplotlib==3.5.1
mysqlclient
pytest
seaborn